refactor: fix bugs and make UT works
